% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaHelper.R
\name{adjustBaseline}
\alias{adjustBaseline}
\title{Adjust for baseline differences using Centered-Log Ratio}
\usage{
adjustBaseline(se_object = NULL)
}
\arguments{
\item{se_object}{SummarizedExperiment object contains omics count/level matrix 
and  metadata}
}
\value{
a SummarizedExperiment object with the adjusted baseline assay
}
\description{
Adjust for baseline differences using Centered-Log Ratio
}
\examples{
data(omicslonda_data_example)
adjusted_se = adjustBaseline(omicslonda_data_example$omicslonda_se_object)
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
