% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveFitting.R
\name{curveFitting}
\alias{curveFitting}
\title{Fit longitudinal data}
\usage{
curveFitting(formula = Count ~ Time, df, method = "ssnbinomial", points)
}
\arguments{
\item{formula}{formula to be passed to the regression model}

\item{df}{dataframe has the Count, Group, Subject, Time}

\item{method}{fitting method (ssgaussian)}

\item{points}{points at which the prediction should happen}
}
\value{
returns the fitted model
}
\description{
Fits longitudinal samples from the same group using negative binomial smoothing splines or LOWESS
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
