% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diff_simulatedDataset_norm}
\alias{diff_simulatedDataset_norm}
\title{Simulated dataset}
\format{A data frame with 1000 simulated features}
\usage{
diff_simulatedDataset_norm
}
\description{
The dataset is used for testing OmicsLonDA
}
\keyword{datasets}
