% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslonda.R
\name{omicslonda}
\alias{omicslonda}
\title{Omics Longitudinal Differential Analysis for one feature}
\usage{
omicslonda(formula = Count ~ Time, df, n.perm = 500,
  fit.method = "ssgaussian", points, text = 0, parall = FALSE,
  pvalue.threshold = 0.05, adjust.method = "BH", time.unit = "days",
  ylabel = "Normalized Count", col = c("blue", "firebrick"),
  prefix = "Test", DrawTestStatDist = FALSE)
}
\arguments{
\item{formula}{formula to be passed to the regression model}

\item{df}{dataframe that contains (subject, time, group, count,
normalizedCount, or any any other covariates) for each feature}

\item{n.perm}{number of permutations.}

\item{fit.method}{fitting method (ssguassian).}

\item{points}{points at which the prediction should happen.}

\item{text}{Feature's name.}

\item{parall}{boolean to indicate whether to use multicore.}

\item{pvalue.threshold}{p-value threshold cutoff for identifing significant
time intervals.}

\item{adjust.method}{multiple testing correction method.}

\item{time.unit}{time unit used in the Time vector (hours, days, weeks,
months, etc.)}

\item{ylabel}{text to be shown on the y-axis of all generated figures
(default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}

\item{DrawTestStatDist}{boolean to indicate if the histogram of the testStat
needs to be plotted or not. Default is "FALSE" since the histogram is usually
a big file}
}
\value{
returns a list of the significant time intervals for the tested
feature.
}
\description{
Find significant time intervals of omic feature
}
\examples{
data(diff_simulatedDataset_norm)
points = seq(1, 200, length.out = 200)
output.omicslonda_diff_1 = omicslonda(formula = normalizedCount ~ Time, 
			      df = diff_simulatedDataset_norm[[1]], n.perm = 10, 
                                     fit.method = "ssgaussian", points = points,
                                     text = "sim_f1", parall = FALSE, pvalue.threshold = 0.05,
                                     adjust.method = "BH", col = c("blue", "green"),
                                     prefix = "OmicsLonDA_clr_f1", ylabel = "CLR-NormalizedCount",
                                     DrawTestStatDist = FALSE, time.unit = "days")
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
