% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{omicslonda_data_example}
\alias{omicslonda_data_example}
\title{Simulated dataset}
\format{A data frame with 10 simulated features}
\usage{
omicslonda_data_example
}
\description{
Datasets used for examples/testing OmicsLonDA
}
\keyword{datasets}
