% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\name{permutationMC2}
\alias{permutationMC2}
\title{Permute group labels}
\usage{
permutationMC2(formula = Count ~ Time, perm.dat, n.perm = 500,
  method = "ssgaussian", points, parall = FALSE, prefix)
}
\arguments{
\item{formula}{formula to be passed to the regression model}

\item{perm.dat}{dataframe has the Count, Group, Subject, Time}

\item{n.perm}{number of permutations}

\item{method}{The fitting method (ssgaussian)}

\item{points}{The points at which the prediction should happen}

\item{parall}{boolean to indicate whether to use multicore.}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
returns the fitted model for all the permutations
}
\description{
Permutes the group label of the samples in order to construct the testStatistics empirical distibution
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
