% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeFeature}
\alias{visualizeFeature}
\title{Visualize Longitudinal Feature}
\usage{
visualizeFeature(df, text, group.levels, unit = "days",
  ylabel = "Normalized Count", col = c("blue", "firebrick"),
  prefix = "Test")
}
\arguments{
\item{df}{dataframe has the Count, Group, ID, Time}

\item{text}{feature name}

\item{group.levels}{The two level's name}

\item{unit}{time interval unit}

\item{ylabel}{text to be shown on the y-axis of all generated figures (default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\description{
Visualize Longitudinal Feature
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
