% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeFeatureSpline}
\alias{visualizeFeatureSpline}
\title{Visualize the feature trajectory with the fitted Splines}
\usage{
visualizeFeatureSpline(se_object = NULL, omicslonda_object = NULL,
  method = "ssgaussian", text = "FeatureName", unit = "days",
  ylabel = "Normalized Count", col = c("blue", "firebrick"),
  prefix = "Test")
}
\arguments{
\item{se_object}{SummarizedExperiment object contains omics count/level matrix 
and  metadata}

\item{omicslonda_object}{The returned object from omicslonda analysis}

\item{method}{The fitting method (ssgaussian)}

\item{text}{feature name}

\item{unit}{time unit used in Time vector (hours, days, weeks, months, etc.)}

\item{ylabel}{text to be shown on the y-axis of all generated figures
(default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
null
}
\description{
Plot the longitudinal features along with the fitted splines
}
\examples{
data(omicslonda_data_example)
visualizeFeatureSpline(se_object = omicslonda_data_example$omicslonda_se_object, 
                      omicslonda_object = omicslonda_data_example$omicslonda_results, 
                      method = "ssgaussian", text = "Feature_1", unit = "days",
                      ylabel = "Normalized Count", col = c("blue", "firebrick"),
                      prefix = "OmicsLonDA_example")
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
