% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeTestStatHistogram}
\alias{visualizeTestStatHistogram}
\title{Visualize test statistics empirical distribution}
\usage{
visualizeTestStatHistogram(omicslonda_object = NULL,
  text = "FeatureName", method = "ssgaussian", prefix = "Test")
}
\arguments{
\item{omicslonda_object}{The returned object from omicslonda analysis}

\item{text}{Feature name}

\item{method}{fitting method}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
null
}
\description{
Visualize sest statistics empirical distribution
}
\examples{
library(SummarizedExperiment)
data(omicslonda_data_example)
omicslonda_data_example$ome_matrix[1:5, 1:5]
se_ome_matrix = as.matrix(omicslonda_data_example$ome_matrix)
se_metadata = DataFrame(omicslonda_data_example$metadata)
omicslonda_se_object = SummarizedExperiment(assays=list(se_ome_matrix),
                                            colData = se_metadata)
omicslonda_se_object_adjusted = adjustBaseline(se_object = omicslonda_se_object)
omicslonda_test_object = omicslonda_se_object_adjusted[1,]
points = seq(1, 500, length.out = 500)
res = omicslonda(se_object = omicslonda_test_object, n.perm = 10,
                 fit.method = "ssgaussian", points = points, text = "Feature_1",
                 parall = FALSE, pvalue.threshold = 0.05, 
                 adjust.method = "BH", time.unit = "days",
                 ylabel = "Normalized Count",
                 col = c("blue", "firebrick"), prefix = "OmicsLonDA_example")
visualizeTestStatHistogram(omicslonda_object = res, text = "Feature_1", 
                 method = "ssgaussian", prefix = "OmicsLonDA_example")
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
