% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaHelper.R
\name{findSigInterval}
\alias{findSigInterval}
\title{Find Significant Interval based on testStat}
\usage{
findSigInterval(adjusted.pvalue, threshold = 0.05, sign)
}
\arguments{
\item{adjusted.pvalue}{vector of the adjusted p-value}

\item{threshold}{p-value cut off}

\item{sign}{vector hold area sign of each time interval}
}
\value{
returns a list of the start and end points of all significant
time intervals
}
\description{
Find Significant Interval based on testStat
}
\examples{
library(SummarizedExperiment)
padjusted = abs(rnorm(10, mean = 0.05, sd = 0.04))
sign = sample(x = c(1,-1), 10, replace = TRUE)
intervals = findSigInterval(adjusted.pvalue = padjusted, 
                        threshold = 0.05, sign = sign)
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
