% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslonda.R
\name{omicslonda}
\alias{omicslonda}
\title{Omics Longitudinal Differential Analysis for one feature}
\usage{
omicslonda(se_object = NULL, n.perm = 500, fit.method = "ssgaussian",
  points = NULL, text = "FeatureName", parall = FALSE,
  pvalue.threshold = 0.05, adjust.method = "BH", time.unit = "days",
  ylabel = "Normalized Count", col = c("blue", "firebrick"),
  prefix = "Test")
}
\arguments{
\item{se_object}{SummarizedExperiment object contains omics count/level matrix 
and  metadata contains (subject, time, group, and any any other covariates)}

\item{n.perm}{number of permutations.}

\item{fit.method}{fitting method (ssguassian).}

\item{points}{points at which the prediction should happen.}

\item{text}{Feature's name.}

\item{parall}{boolean to indicate whether to use multicore.}

\item{pvalue.threshold}{p-value threshold cutoff for identifing significant
time intervals.}

\item{adjust.method}{multiple testing correction method.}

\item{time.unit}{time unit used in the Time vector (hours, days, weeks,
months, etc.)}

\item{ylabel}{text to be shown on the y-axis of all generated figures
(default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
a list of the significant time intervals for the tested
feature, fitted model for each group, null distribution of the test statistic 
of the tested feature, and the original input data.
}
\description{
Find significant time intervals of omic feature
}
\examples{
library(SummarizedExperiment)
data(omicslonda_data_example)
omicslonda_se_object_adjusted = adjustBaseline(
                 se_object = omicslonda_data_example$omicslonda_se_object)
omicslonda_test_object = omicslonda_se_object_adjusted[1,]
points = seq(1, 500, length.out = 500)
res = omicslonda(se_object = omicslonda_test_object, n.perm = 10,
                 fit.method = "ssgaussian", points = points, text = "Feature_1",
                 parall = FALSE, pvalue.threshold = 0.05, 
                 adjust.method = "BH", time.unit = "days",
                 ylabel = "Normalized Count",
                 col = c("blue", "firebrick"), prefix = tempfile())
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
