% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaMCPermutation.R
\name{permutationMC2}
\alias{permutationMC2}
\title{Permute group labels}
\usage{
permutationMC2(formula = Count ~ Time, perm.dat, n.perm = 500,
  method = "ssgaussian", points, parall = FALSE, prefix)
}
\arguments{
\item{formula}{formula to be passed to the regression model}

\item{perm.dat}{dataframe has the Count, Group, Subject, Time}

\item{n.perm}{number of permutations}

\item{method}{The fitting method (ssgaussian)}

\item{points}{The points at which the prediction should happen}

\item{parall}{boolean to indicate whether to use multicore.}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
returns the fitted model for all the permutations
}
\description{
Permutes the group label of the samples in order to construct the
testStatistics empirical distibution
}
\examples{
data(diff_simulatedDataset_norm)
df = diff_simulatedDataset_norm[[1]]
Group = as.character(df$Group)
group.levels = sort(unique(Group))
gr.1 = as.character(group.levels[1])
gr.2 = as.character(group.levels[2])
levels(df$Group) = c(levels(df$Group), "0", "1")
df$Group[which(df$Group == gr.1)] = 0
df$Group[which(df$Group == gr.2)] = 1
group.0 = df[df$Group == 0, ]
group.1 = df[df$Group == 1, ]
points = seq(100, 130)
model = curveFitting(formula = Count ~ Time, df, method= "ssgaussian", points)
perm  = permutationMC2(formula = Count ~ Time, perm.dat = df, n.perm = 10,
method = "ssgaussian", points = points, parall = FALSE, prefix = "Test")
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
