% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaHelper.R
\name{testStat}
\alias{testStat}
\title{Calculate Test-Statistic of each feature's time interval}
\usage{
testStat(curve.fit.df)
}
\arguments{
\item{curve.fit.df}{gss data object of the fitted spline}
}
\value{
returns testStat for all time intervals
}
\description{
Calculate Test-Statistic of each feature's time interval
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
