% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaHelper.R
\name{testStat}
\alias{testStat}
\title{Calculate Test-Statistic of each feature's time interval}
\usage{
testStat(curve.fit.df)
}
\arguments{
\item{curve.fit.df}{gss data object of the fitted spline}
}
\value{
a list that has the test statistic for each time interval
}
\description{
Calculate Test-Statistic of each feature's time interval
}
\examples{
library(SummarizedExperiment)
data("omicslonda_data_example")
model = omicslonda_data_example$omicslonda_results$model
stat = testStat(model)$testStat
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
