% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveFitting.R
\name{testStatPermutation}
\alias{testStatPermutation}
\title{Calculate testStat of each feature's time interval for all permutations}
\usage{
testStatPermutation(perm)
}
\arguments{
\item{perm}{list has all the permutated models}
}
\value{
returns a list of all permutation area ratio
}
\description{
Calculate testStat of each feature's time interval for all permutations
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
