% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaHelper.R
\name{testStatPermutation}
\alias{testStatPermutation}
\title{Calculate testStat of each feature's time interval for all permutations}
\usage{
testStatPermutation(perm)
}
\arguments{
\item{perm}{list has all the permutated models}
}
\value{
returns a list of all permutation area ratio
}
\description{
Calculate testStat of each feature's time interval for all permutations
}
\examples{
data(diff_simulatedDataset_norm)
df = diff_simulatedDataset_norm[[1]]
Group = as.character(df$Group)
group.levels = sort(unique(Group))
gr.1 = as.character(group.levels[1])
gr.2 = as.character(group.levels[2])
levels(df$Group) = c(levels(df$Group), "0", "1")
df$Group[which(df$Group == gr.1)] = 0
df$Group[which(df$Group == gr.2)] = 1
group.0 = df[df$Group == 0, ]
group.1 = df[df$Group == 1, ]
points = seq(100, 130)
model = curveFitting(formula = Count ~ Time, df, method= "ssgaussian", 
        points)
perm  = permutationMC2(formula = Count ~ Time, perm.dat = df, n.perm = 10,
           method = "ssgaussian", points = points, parall = FALSE,
           prefix = "Test")
test.stat.prem = testStatPermutation(perm)
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
