% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeArea}
\alias{visualizeArea}
\title{Visualize significant time interval}
\usage{
visualizeArea(formula = Count ~ Time, model.ss, method, start, end, text,
  group.levels, unit = "days", ylabel = "Normalized Count",
  col = c("blue", "firebrick"), prefix = "Test")
}
\arguments{
\item{formula}{formula used in method}

\item{model.ss}{The fitted model}

\item{method}{Fitting method (ssgaussian)}

\item{start}{Vector of the start points of the time intervals}

\item{end}{Vector of the end points of the time intervals}

\item{text}{Feature name}

\item{group.levels}{Level's name}

\item{unit}{time unit used in the Time vector (hours, days, weeks, months, etc.)}

\item{ylabel}{text to be shown on the y-axis of all generated figures (default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
null
}
\description{
Visualize significant time interval
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
