% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeArea}
\alias{visualizeArea}
\title{Visualize significant time interval}
\usage{
visualizeArea(omicslonda_object = NULL, fit.method = "ssgaussian",
  text = "FeatureName", unit = "days", ylabel = "Normalized Count",
  col = c("blue", "firebrick"), prefix = "Test")
}
\arguments{
\item{omicslonda_object}{The returned object from omicslonda analysis}

\item{fit.method}{Fitting method (ssgaussian)}

\item{text}{Feature name}

\item{unit}{time unit used in the Time vector
(hours, days, weeks, months, etc.)}

\item{ylabel}{text to be shown on the y-axis of all generated figures
(default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
null
}
\description{
Visualize significant time interval
}
\examples{
library(SummarizedExperiment)
data(omicslonda_data_example)
visualizeArea(omicslonda_object = omicslonda_data_example$omicslonda_results, 
                 fit.method = "ssgaussian",
                 text = "Feature_1", unit = "days", 
                 ylabel = "Normalized Count", col =
                 c("blue", "firebrick"), prefix = tempfile())
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
