% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeFeature}
\alias{visualizeFeature}
\title{Visualize Longitudinal Feature}
\usage{
visualizeFeature(formula = Count ~ Time, df, text, group.levels,
  unit = "days", ylabel = "Normalized Count", col = c("blue",
  "firebrick"), prefix = "Test")
}
\arguments{
\item{formula}{formula used in method}

\item{df}{dataframe has the Count, Group, ID, Time}

\item{text}{feature name}

\item{group.levels}{The two level's name}

\item{unit}{time interval unit}

\item{ylabel}{text to be shown on the y-axis of all generated figures
(default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
null
}
\description{
Visualize Longitudinal Feature
}
\examples{
data(diff_simulatedDataset_norm)
df = diff_simulatedDataset_norm[[1]]
Group = as.character(df$Group)
group.levels = sort(unique(Group))
gr.1 = as.character(group.levels[1])
gr.2 = as.character(group.levels[2])
levels(df$Group) = c(levels(df$Group), "0", "1")
df$Group[which(df$Group == gr.1)] = 0
df$Group[which(df$Group == gr.2)] = 1
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
