% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeFeature}
\alias{visualizeFeature}
\title{Visualize Longitudinal Feature}
\usage{
visualizeFeature(se_object = NULL, text = "featureName",
  unit = "days", ylabel = "Normalized Count", col = c("blue",
  "firebrick"), prefix = "Test")
}
\arguments{
\item{se_object}{SummarizedExperiment object contains omics count/level matrix 
and  metadata}

\item{text}{feature name}

\item{unit}{time interval unit}

\item{ylabel}{text to be shown on the y-axis of all generated figures
(default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
null
}
\description{
Visualize Longitudinal Feature
}
\examples{
library(SummarizedExperiment)
data("omicslonda_data_example")
omicslonda_se_object_adjusted = adjustBaseline(
                 se_object = omicslonda_data_example$omicslonda_se_object)
omicslonda_test_object = omicslonda_se_object_adjusted[1,]
visualizeFeature(se_object = omicslonda_test_object, text = "Feature_1",
                 unit = "days", ylabel = "Normalized Count", 
                 col = c("blue", "firebrick"), prefix = tempfile())
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
