% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeFeatureSpline2}
\alias{visualizeFeatureSpline2}
\title{Visualize the feature trajectory with the fitted Splines}
\usage{
visualizeFeatureSpline2(formula = Count ~ Time, df, model, method, text,
  group.levels, unit = "days", ylabel = "Normalized Count",
  col = c("blue", "firebrick"), prefix = "Test")
}
\arguments{
\item{formula}{formula to be passed to the regression model}

\item{df}{dataframe has the Count , Group, ID, Time}

\item{model}{the fitted model}

\item{method}{The fitting method (ssgaussian)}

\item{text}{feature name}

\item{group.levels}{The two level's name}

\item{unit}{time unit used in Time vector (hours, days, weeks, months, etc.)}

\item{ylabel}{text to be shown on the y-axis of all generated figures
(default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
null
}
\description{
Plot the longitudinal features along with the fitted splines
}
\examples{
data(diff_simulatedDataset_norm)
df = diff_simulatedDataset_norm[[1]]
Group = as.character(df$Group)
group.levels = sort(unique(Group))
gr.1 = as.character(group.levels[1])
gr.2 = as.character(group.levels[2])
levels(df$Group) = c(levels(df$Group), "0", "1")
df$Group[which(df$Group == gr.1)] = 0
df$Group[which(df$Group == gr.2)] = 1
group.0 = df[df$Group == 0, ]
group.1 = df[df$Group == 1, ]
points = seq(100, 130)
model = curveFitting(formula = Count ~ Time, df, method= "ssgaussian", 
        points)
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
