% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeTestStatHistogram}
\alias{visualizeTestStatHistogram}
\title{Visualize Test Statistics empirical distribution}
\usage{
visualizeTestStatHistogram(permuted, text, method, prefix = "Test", modelStat)
}
\arguments{
\item{permuted}{Permutation of the permuted data}

\item{text}{Feature name}

\item{method}{fitting method}

\item{prefix}{prefix to be used to create directory for the analysis results}

\item{modelStat}{test statistics}
}
\description{
Visualize Test Statistics empirical distribution for each time interval
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
