% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicslondaVisualization.R
\name{visualizeTestStatHistogram}
\alias{visualizeTestStatHistogram}
\title{Visualize test statistics empirical distribution}
\usage{
visualizeTestStatHistogram(omicslonda_object = NULL,
  text = "FeatureName", fit.method = "ssgaussian", prefix = "Test")
}
\arguments{
\item{omicslonda_object}{The returned object from omicslonda analysis}

\item{text}{Feature name}

\item{fit.method}{fitting method}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\value{
null
}
\description{
Visualize sest statistics empirical distribution
}
\examples{
library(SummarizedExperiment)
data(omicslonda_data_example)
visualizeTestStatHistogram(omicslonda_object = omicslonda_data_example$omicslonda_results, 
                 text = "Feature_1", 
                 fit.method = "ssgaussian", prefix = tempfile())
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
